ai-service/pyproject.toml
[project]
name = "ai-service"
version = "0.1.0"
description = "FastAPI microservice for HydraAI Inspect"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "pydantic",
  "pydantic-settings",
  "uvicorn",
  "numpy",
  "torch",
  "ultralytics",
  "pillow",
  "python-multipart"
]

[project.optional-dependencies]
dev = ["pytest", "httpx"]

[tool.uvicorn]
factory = false
app = "src.main:app"
